refactor: migrate TerminalRenderer tests to renderer module
Split terminal_tests.rs: TerminalRenderer tests moved to
src/renderer/terminal_tests.rs, TerminalAdapter and TerminalPlugin
tests remain in src/im/terminal_tests.rs.

Source: user
Type: refactor
